% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateMonthly.R
\name{aggregateMonthly}
\alias{aggregateMonthly}
\title{This function takes a dataframe with weekly precipitation and concentration values,}
\usage{
aggregateMonthly(bi, site, siObs, obs, obsi, totT, strtYrMo)
}
\description{
This function takes a dataframe with weekly precipitation and concentration values,
}
\keyword{internal}
