% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataIn.R
\docType{data}
\name{defaultInput}
\alias{defaultInput}
\title{Default input for getCov()}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(defaultInput)
}
\description{
Default input for getCov()
}
\examples{
data(defaultInput)
df <- defaultInput
df 
}
\keyword{datasets}
