% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data50.R
\docType{data}
\name{dfRes50}
\alias{dfRes50}
\title{Instance of dataframe of residuals for 50 sites
Sites included:
For dates:
For compound:
Missing values filled in use fitted LambertW distribution}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(dfRes50)
}
\description{
Instance of dataframe of residuals for 50 sites
Sites included:
For dates:
For compound:
Missing values filled in use fitted LambertW distribution
}
\examples{
data(dfRes50)
dfRes <- dfRes50
dfRes
}
\keyword{datasets}
