% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCovL.R
\name{getCovL}
\alias{getCovL}
\title{creates covariance matrix from normalized NADP monitor data
gamma version with optional inputs}
\usage{
getCovL(df)
}
\arguments{
\item{df}{data frame of input}
}
\value{
List of model summaries at each site, covariance matrix and plots if inputted as T
}
\description{
creates covariance matrix from normalized NADP monitor data
gamma version with optional inputs
}
\examples{
  getCovL(structure(list(
weeklyB = FALSE, startdateStr = "01/01/83 00:00", enddateStr   = "12/31/86 00:00", 
comp = "SO4", use36 = TRUE, siteAdd = NULL, outlierDatesbySite = NULL,
siteOutliers = NULL,  removeOutliers = NULL, plotMulti = FALSE,  sitePlot = NULL,
plotAll = FALSE, writeMat = FALSE, seas = 12, r = 1, k = 1),
.Names = c("weeklyB","startdateStr","enddateStr","comp","use36","siteAdd",
"outlierDatesbySite","siteOutliers","removeOutliers","plotMulti","sitePlot",
"plotAll","writeMat","seas","r","k")  ,row.names = c(NA, -1L) 
,class = "data.frame"))
}
