% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSites.R
\name{getSites}
\alias{getSites}
\title{returns a vector of siteID's of sites with the most data for the specified time period for specific compunds or pH}
\usage{
getSites(startdateStr, enddateStr, maxn, mins, comp, optR)
}
\arguments{
\item{startdateStr}{String, when to start analyzing data, format = "m/d/y H:M"}

\item{enddateStr}{String, when to stop analyzing data, format = "m/d/y H:M"}

\item{maxn}{Maximum number of sites required}

\item{mins}{Minimum number of data required for each site}

\item{comp}{Vector of strings, strings of compounds}

\item{optR}{String, option of region "N","NE","S","W",""}
}
\value{
Returns list of sites that have the largest amount of data for the given dates and compounds
}
\description{
returns a vector of siteID's of sites with the most data for the specified time period for specific compunds or pH
}
\examples{
getSites("01/01/83 00:00","12/31/86 00:00",30,102,"SO4","")
}
