% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contain.R
\name{\%contain\%}
\alias{\%contain\%}
\title{code posted by Marat Talipov in stackoverflow post link: https://stackoverflow.com/questions/34445106/test-if-vector-is-contained-in-another-vector-including-repetitions}
\usage{
values \%contain\% x
}
\description{
code posted by Marat Talipov in stackoverflow post link: https://stackoverflow.com/questions/34445106/test-if-vector-is-contained-in-another-vector-including-repetitions
}
\keyword{internal}
