% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independenceTest.R
\name{independenceTest}
\alias{independenceTest}
\title{returns a vector of siteID's of sites with the most data for the specified time period for specific compunds or pH}
\usage{
independenceTest(dfRes)
}
\arguments{
\item{dfRes}{Data frame od residuals}
}
\value{
Returns result of independence test and correlation matrix
}
\description{
returns a vector of siteID's of sites with the most data for the specified time period for specific compunds or pH
}
\examples{
independenceTest(data.frame(matrix(runif(500, min= -1, max =1), ncol=10)))
}
