% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambertWfit.R
\name{lambertWfit}
\alias{lambertWfit}
\title{returns LambertW transformed residuals}
\usage{
lambertWfit(dfRes)
}
\arguments{
\item{dfRes}{dataframe,}
}
\value{
Returns a dataframe of LambertW transformed residuals
}
\description{
returns LambertW transformed residuals
}
\examples{
lambertWfit(data.frame(matrix(runif(n = 100), ncol = 5 )))
}
\keyword{internal}
