% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambertWtransform.R
\name{lambertWtransform}
\alias{lambertWtransform}
\title{returns LambertW transformed residuals}
\usage{
lambertWtransform(dfRes, plotMulti, writeMat)
}
\arguments{
\item{dfRes}{dataframe,}

\item{plotMulti}{Binary,}

\item{writeMat}{Binary,}
}
\value{
Returns a dataframe of LambertW transformed residuals
}
\description{
returns LambertW transformed residuals
}
\examples{
lambertWtransform(data.frame(matrix(runif(n = 100), ncol = 5 )),plotMulti = FALSE, writeMat = FALSE)
}
