% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMax1.R
\docType{data}
\name{maxd1Cov}
\alias{maxd1Cov}
\title{Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/86 00:00","12/31/94 00:00",50,200,"SO4",1)}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data("maxd1Cov")
}
\description{
Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/86 00:00","12/31/94 00:00",50,200,"SO4",1)
}
\examples{

data("maxd1Cov")
cov1 <- maxd1Cov
cov1
}
\keyword{datasets}
