% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMax2.R
\docType{data}
\name{maxd2Cov}
\alias{maxd2Cov}
\title{Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"SO4",1)}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data("maxd2Cov")
}
\description{
Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"SO4",1)
}
\examples{

data("maxd2Cov")
cov2 <- maxd2Cov
cov2
}
\keyword{datasets}
