% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMax3.R
\docType{data}
\name{maxd3Cov}
\alias{maxd3Cov}
\title{Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"NO3",1)}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data("maxd3Cov")
}
\description{
Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"NO3",1)
}
\examples{

data("maxd3Cov")
cov3 <- maxd3Cov
cov3
}
\keyword{datasets}
