% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMax4.R
\docType{data}
\name{maxd4Cov}
\alias{maxd4Cov}
\title{Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"Na",1) 
Covariance produced after applying function lambertWtransform on residualDataNA output from getCov()}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data("maxd4Cov")
}
\description{
Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"Na",1) 
Covariance produced after applying function lambertWtransform on residualDataNA output from getCov()
}
\examples{

data("maxd4Cov")
cov4 <- maxd4Cov
cov4
}
\keyword{datasets}
