% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataMax5.R
\docType{data}
\name{maxd5Cov}
\alias{maxd5Cov}
\title{Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"ph",1)}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data("maxd5Cov")
}
\description{
Instance of covariance matrix for 50 sites
Site list produced by maxDistSites("01/01/07 00:00","12/31/14 00:00",50,230,"ph",1)
}
\examples{

data("maxd5Cov")
cov5 <- maxd5Cov
cov5
}
\keyword{datasets}
