% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxDistSites.R
\name{maxDistSites}
\alias{maxDistSites}
\title{Returns a vector of siteID's of sites with the most data for the specified time period for specific chemicals or pH}
\usage{
maxDistSites(startdateStr, enddateStr, maxn, mins, comp, startingSite)
}
\arguments{
\item{startdateStr}{String, when to start analyzing data, format = "m/d/y H:M"}

\item{enddateStr}{String, when to stop analyzing data, format = "m/d/y H:M"}

\item{maxn}{Number of sites required}

\item{mins}{Minimum number of data required for each site}

\item{comp}{Vector of strings, strings of compounds}

\item{startingSite}{Integer, largest sample index of initial site to include}
}
\value{
Returns list of sites that have the largest amount of data for the given dates and pollutants
}
\description{
Returns a vector of siteID's of sites with the most data for the specified time period for specific chemicals or pH
}
\examples{
maxDistSites("01/01/83 00:00","12/31/86 00:00",50,100,"SO4",1)
}
