% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataG.R
\docType{data}
\name{NADPgeo}
\alias{NADPgeo}
\title{NADP monitor GPS location for all sites}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(NADPgeo)
}
\description{
Location data for all sites currently or formerly active in the NTN network.
}
\examples{
data(NADPgeo)
 Latitude   <- attr(NADPgeo, "Latitude (DD)")
siteID <- NADPgeo$`Site ID`
}
\references{
National Atmospheric Deposition Program (NRSP-3). 2019. NADP Program Office, Wisconsin State Laboratory of Hygiene,
465 Henry Mall, Madison, WI 53706.
(\href{http://nadp.slh.wisc.edu/data/NTN/ntnAllsites.aspx}{NADP/NTN})
}
\keyword{datasets}
