% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataP.R
\docType{data}
\name{preDailyCSV}
\alias{preDailyCSV}
\title{Daily NADP precipitation data}
\format{An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.}
\usage{
data(preDailyCSV)
}
\description{
Daily NADP data of percipitation depth in inches for all sites from 1979 to 2017.
}
\examples{
data(preDailyCSV)
 amount   <- attr(preDailyCSV, "amount")
siteID <- preDailyCSV$siteID
}
\references{
National Atmospheric Deposition Program (NRSP-3). 2019. NADP Program Office, Wisconsin State Laboratory of Hygiene,
465 Henry Mall, Madison, WI 53706.
(\href{http://nadp.slh.wisc.edu/data/NTN/ntnAllsites.aspx}{NADP/NTN})
}
\keyword{datasets}
