% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reEvaluateSites.R
\name{reEvaluateSites}
\alias{reEvaluateSites}
\title{analyzes data without outliers}
\usage{
reEvaluateSites(dfInp, tafNA, startdate, enddate, totT, y0, ylog, e, e2,
  mods, vpredl, outlierDatesbySite, outSites, cati, strtYrMo, endYrMo)
}
\description{
analyzes data without outliers
}
\keyword{internal}
