% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/takeOutAllOutliers.R
\name{takeOutAllOutliers}
\alias{takeOutAllOutliers}
\title{returns LambertW transformed residuals}
\usage{
takeOutAllOutliers(sites, rosnerResult, cati)
}
\arguments{
\item{dfRes}{dataframe,}

\item{plotMulti}{}
}
\value{
Returns a dataframe of LambertW transformed residuals
}
\description{
returns LambertW transformed residuals
}
\keyword{internal}
