% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weeklyConc.R
\name{weeklyConc}
\alias{weeklyConc}
\title{gets weekly values and concentrations from daily precipitayion data}
\usage{
weeklyConc(binary, site, siObs, obs, obsi, startdate)
}
\arguments{
\item{binary}{Is pH being analyzed on the current loop? T/F (generated automatically in loop)}

\item{site}{Data with compund/pH weekly values}

\item{siObs}{Number of compunds analyzed}

\item{obsi}{Index of current compund being analyzed}

\item{startdate}{Startdate of analysis}
}
\value{
sitew data that has weekly concentrations of compound, in the case of pH it gives exactly the weekly value plus a column with weekly precipitation values
}
\description{
gets weekly values and concentrations from daily precipitayion data
}
\keyword{internal}
