% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataW.R
\docType{data}
\name{weeklyCSV}
\alias{weeklyCSV}
\title{Weekly NADP data}
\format{An object of class \code{data.frame} with 424871 rows and 31 columns.}
\source{
\href{http://nadp.slh.wisc.edu/data/NTN/ntnAllsites.aspx}{NADP/NTN}
}
\usage{
data(weeklyCSV)
}
\description{
Weekly NADP data of 10 deposits measured in mg/L for all sites from 1978 to January 2019.
}
\examples{
data(weeklyCSV)
 SO4   <- attr(weeklyCSV, "SO4")
siteID <- weeklyCSV$siteID
}
\references{
National Atmospheric Deposition Program (NRSP-3). 2019. NADP Program Office, Wisconsin State Laboratory of Hygiene,
465 Henry Mall, Madison, WI 53706.
(\href{http://nadp.slh.wisc.edu/data/NTN/ntnAllsites.aspx}{NADP/NTN})
}
\keyword{datasets}
