% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov}
\alias{getCov}
\title{Creates covariance matrix from normalized NADP monitor data
gamma version with optional inputs}
\usage{
getCov(df)
}
\arguments{
\item{df}{data frame of input arguments specified in vignette and default set of inputs are in data("defultInput")}
}
\value{
List of model summaries at each site, covariance matrix, sites analyzed, predicted values by univariate models, data frame of residuals and more (see vignette for full list)
}
\description{
Creates covariance matrix from normalized NADP monitor data
gamma version with optional inputs
}
\examples{
getCov(structure(list(startdateStr = "01/01/83 00:00", enddateStr   = "12/31/86 00:00", 
comp = "SO4", use36 = TRUE, siteAdd = NULL, outlierDatesbySite = NULL,
siteOutliers = NULL,  removeOutliers = NULL, plotMulti = FALSE,  sitePlot = NULL,
plotAll = FALSE, writeMat = FALSE, seas = 12, r = 1, k = 1),
.Names = c("startdateStr","enddateStr","comp","use36","siteAdd",
"outlierDatesbySite","siteOutliers","removeOutliers","plotMulti","sitePlot",
"plotAll","writeMat","seas","r","k")  ,row.names = c(NA, -1L) 
,class = "data.frame"))
}
