% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCov.R
\name{getCov}
\alias{getCov}
\title{creates covariance matrix from normalized NADP monitor data
gamma version with optional inputs}
\usage{
getCov(df)
}
\arguments{
\item{df}{data frame of input}
}
\value{
List of model summaries at each site, covariance matrix and plots if inputted as T
}
\description{
creates covariance matrix from normalized NADP monitor data
gamma version with optional inputs
}
\examples{
  getCov(structure(list(
weeklyB = FALSE, startdateStr = "01/01/83 00:00", enddateStr   = "12/31/86 00:00", 
use36 = TRUE, siteAdd = NULL, outlierDatesbySite = NULL, showOutliers = FALSE,
siteOutliers = NULL, comp = "SO4", plotMulti = FALSE, plotB = FALSE, sitePlot = NULL,
plotAll = FALSE, writeMat = FALSE, seas = 12, r = 1, k = 1)
,.Names = c("weeklyB","startdateStr","enddateStr","use36","siteAdd",
           "outlierDatesbySite","showOutliers","siteOutliers","comp","plotMulti",
           "plotB","sitePlot","plotAll","writeMat","seas","r","k") 
,row.names = c(NA, -1L) 
,class = "data.frame"))
}
