% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambertWtransform.R
\name{lambertWtransform}
\alias{lambertWtransform}
\title{Returns LambertW transformed residuals}
\usage{
lambertWtransform(dfRes, plotMulti, writeMat)
}
\arguments{
\item{dfRes}{Dataframe of residuals (can take NA values)}

\item{plotMulti}{Binary, TRUE to plot the multivariate qq plot}

\item{writeMat}{Binary, TRUE to write a mat file of the resulting covariance matrix with the name "covSitesLambertW.mat"}
}
\value{
Returns a dataframe of LambertW transformed residuals, multivariate and univariate variables analysis, and the resulting covariance matrix
}
\description{
Returns LambertW transformed residuals
}
\examples{
lambertWtransform(data.frame(matrix(runif(n = 100), ncol = 5 )),plotMulti = FALSE, writeMat = FALSE)
}
