% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambertWtransform.R
\name{lambertWtransform}
\alias{lambertWtransform}
\title{returns LambertW transformed residuals}
\usage{
lambertWtransform(dfRes, plotMulti, writeMat)
}
\arguments{
\item{dfRes}{dataframe, when to start analyzing data, format = "m/d/y H:M"}

\item{plotMulti}{Binary, when to stop analyzing data, format = "m/d/y H:M"}
}
\value{
Returns a dataframe of LambertW transformed residuals
}
\description{
returns LambertW transformed residuals
}
\examples{
lambertWtransform(dfRes = dfRes50,plotMulti = FALSE, writeMat = FALSE)
}
